Active Storage Done

+++ db/schema.rb
@@ -15,8 +15,8 @@ ActiveRecord::Schema.define(version: 2023_04_02_113322) do
   create_table "active_storage_attachments", force: :cascade do |t|
     t.string "name", null: false
     t.string "record_type", null: false
-    t.bigint "record_id", null: false
-    t.bigint "blob_id", null: false
+    t.integer "record_id", null: false
+    t.integer "blob_id", null: false
     t.datetime "created_at", null: false
     t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
     t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
@@ -103,13 +103,13 @@ ActiveRecord::Schema.define(version: 2023_04_02_113322) do
   end
 
 # Could not dump table "transactions" because of following StandardError
-#   Unknown type 'attachment' for column 'documents'
+#   Unknown type 'attachment' for column 'document'
 
 # Could not dump table "user_categories" because of following StandardError
 #   Unknown type 'attachment' for column 'background'
 
 # Could not dump table "users" because of following StandardError
-#   Unknown type 'attachment' for column 'profile_image'
+#   Unknown type 'attachment' for column 'avatar'
 
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
   add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
